future: extracting result marks future as unavailable.
This commit is a follow-up to "future: don't create
temporaries along get_available_state() paths" which
allowed to `get()`/`then()` to be called multiple
times on same `future` instance.

I'm not entirely sure that prohibting such calls is
huge value. This change is kept separately to allow
for dropping it easily during the review.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>